requirements.txt
asgiref==3.7.2
Django==5.0.1
paho-mqtt==1.6.1
psycopg2==2.9.9
sqlparse==0.4.4